Move user, admin components, conexts to its directories, and remove mode switch
