+++ ui/src/common/GoogleSignOutButton.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 import { LogoutRounded } from '@mui/icons-material';
 import { Button, CircularProgress } from '@mui/material';
-import { useSnackbar } from '../SnackbarContext';
 import GoogleLogo from '../assets/google_logo.svg';
 import { useAuth } from '../auth';
+import { useSnackbar } from '../contexts/snackbar';
 
 const GoogleSignOutButton = ({ text }: { text?: string }) => {
   const { logout } = useAuth();
